feat(mespapiers): Konnector UI management without files
Creation of a parent component to `CategoryItemByPaper`
to handle cases where installed konnectors have no files.
