+++ packages/cozy-mespapiers-lib/src/components/Papers/PaperGroup.jsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router-dom'
 
 import List from 'cozy-ui/transpiled/react/List'
 
-import CategoryItemByPaper from './CategoryItemByPaper'
+import { CategoriesByPaper } from './CategoriesByPaper'
 import KonnectorsCategories from './KonnectorsCategories'
 import { useMultiSelection } from '../Hooks/useMultiSelection'
 
@@ -12,8 +12,6 @@ const PaperGroup = ({ papersByCategories, konnectors, selectedThemes }) => {
   const navigate = useNavigate()
   const { isMultiSelectionActive, setSelectedQualificationLabel } =
     useMultiSelection()
-
-  const hasPapers = Object.keys(papersByCategories).length > 0
 
   const goPapersList = qualificationLabel => {
     if (isMultiSelectionActive) {
@@ -25,16 +23,10 @@ const PaperGroup = ({ papersByCategories, konnectors, selectedThemes }) => {
 
   return (
     <List>
-      {hasPapers &&
-        Object.entries(papersByCategories).map(([category, papers], index) => (
-          <CategoryItemByPaper
-            key={category}
-            category={category}
-            papers={papers}
-            isLast={index === Object.entries(papersByCategories).length - 1}
-            onClick={goPapersList}
-          />
-        ))}
+      <CategoriesByPaper
+        papersByCategories={papersByCategories}
+        onClick={goPapersList}
+      />
       <KonnectorsCategories
         konnectors={konnectors}
         selectedThemes={selectedThemes}
